Project wiki support.
This commit contains support for parsing individual wiki
pages along with wiki sitemaps (Confusingly, with the slug
"pages")

It also contains various bug fixes since project wikis are
not git repositories and attempting to use APIs specific to
Git repos will fail with either 403 Forbidden or 404
Repository not found.

Maybe we can could check if a repo is a wiki before
requesting git data from it. That way we can show useful
data to the user using wiki-specific APIs. I might work on
that actually...

Signed-off-by: Leo Gavilieau <xmoo@vern.cc>